♻️ renames Queries folder to SqonBuilder (#1208)
